+++ src/batou/lib/tests/test_buildout.py
@@ -55,9 +55,7 @@ def test_runs_buildout_successfully(root):
         os.path.join(root.environment.workdir_base, "mycomponent/bin/py"))
 
 
-# XFAIL due to https://github.com/flyingcircusio/batou/issues/169
 @pytest.mark.slow
-@pytest.mark.xfail
 @pytest.mark.timeout(60)
 def test_runs_buildout3_successfully(root, output):
     b = Buildout(
@@ -65,6 +63,7 @@ def test_runs_buildout3_successfully(root, output):
         version="3.0.0b1",
         setuptools="54.1.1",
         wheel="0.36.2",
+        pip="21.0.1",
         config=File(
             "buildout3.cfg",
             source=pkg_resources.resource_filename(__name__,