+++ CHANGES.md
@@ -14,6 +14,8 @@
   `./batou migrate`.
   ([#185](https://github.com/flyingcircusio/batou/issues/185))
 
+- Support creating users in MySQL ≥8.
+  ([#242](https://github.com/flyingcircusio/batou/issues/242))
 
 2.3b3 (2021-11-30)
 ------------------