+++ CHANGES.md
@@ -10,12 +10,18 @@
   `batou migrate`.
   ([#185](https://github.com/flyingcircusio/batou/issues/185))
 
+- Allow to pin the `pip` version used in the `Buildout` component.
+  ([#263](https://github.com/flyingcircusio/batou/issues/263))
+
 - Automatically migrate environments and secrets to the new structure using
   `./batou migrate`.
   ([#185](https://github.com/flyingcircusio/batou/issues/185))
 
 - Support creating users in MySQL ≥8.
   ([#242](https://github.com/flyingcircusio/batou/issues/242))
+
+- Allow to check if an `Address` is configured for IPv4 resp. IPv6 using the
+  attributes `require_v4` resp. `require_v6`.
 
 2.3b3 (2021-11-30)
 ------------------